Move Kokkos_OpenMP.hpp to OpenMP/Kokkos_OpenMP.hpp
