+++ core/src/OpenMP/Kokkos_OpenMP_WorkGraphPolicy.hpp
@@ -17,7 +17,7 @@
 #ifndef KOKKOS_OPENMP_WORKGRAPHPOLICY_HPP
 #define KOKKOS_OPENMP_WORKGRAPHPOLICY_HPP
 
-#include <Kokkos_OpenMP.hpp>
+#include <OpenMP/Kokkos_OpenMP.hpp>
 
 namespace Kokkos {
 namespace Impl {
